window icon: set window icon on taskbar to data/ui/logo.png
In UOS system, windows of all kinds of apps have icon that not odds with app name.
So We Use unified icon: logo.png, so that this problem will be solved.
Note: UOS finds Desktop fast links to determine icon of program's window, so set fast links to
anbox-launch.sh is important. And then, anbox-launch.sh will transparent inputs to anbox-session
process.