window: [fix bug]open app in horizontal screen, but window is in vertical state.
Some app opens in horizontal screen, but window on anbox side is in vertical state,
As the result, bottom of the window has black area.
To solve this problem, we adjust window on anbox side when app opens in horizontal
screen state.
But, when window adjusted by user, app follows this adjustment and send
this adjustment to anbox at the second time. So change of this commit has no effect
after window is adjusted by user or 15s timeout.